Add fail() and replace assert(0) and check(..., 1)
If the programmer uses assert(0) and assumes that it will stop the program, this may lead to compiler warnings in release mode. Also, check(..., 1) is kind of a hack. Added a function called fail() which stops the program and reports an error message with line number, and which G++ knows will always stop the program so it deosn't give warnings.