No more bs.
Okay, we're done with this algebra bs. We're just going to do this the old
fashioned way and load everything into a SQLite databes.  Removes counters and
everything like that. Adds the SQLite protocol to be used for keeping track of
the keys. (In theory we could now benchmark sqlite, if we wanted to raise our
databases self esteem.