Unlock the buf in snapshot as soon as possible
When acquiring the first block after calling transaction_t::snapshot(),
the we have to lock on that block (to ensure correct ordering with
respect to any possible write working on that block).
However, after we acquired the lock and generated the version id for the
snapshot, we can release the lock.