Adds the proper UI for import adds import throttling.
The UI is now:
rethinkdb import --memcached-file foo
This is done in a kind of hacky-sack way in that import is just an alias
for serve at this point. There's actually nothing to stop you from
doing:
rethinkdb serve --memcached-file bar
which would have the same effect. This is kind of a pain to change
because I need to decouple some argument parsing things. We'll do it
eventually.

Also throttling is now enabled for import because otherwise it may cause
RethinkDb to go out of memory. With a back-of-thin-air calculation we
figured that 1000 is about the right number to throttle to.