os_signal_monitor is now not a global.
This is because it caused its destructor to be called which was making
us trip some asserts. It's now a class that must be manually
constructed and sets itself up in a global variable when constructed.
It's still a singleton but will only trip at run time if more than 1 has
been constructed. This isn't great it should actually be part of the
thread pool.. but that would require alot of code to be rearranged.