Adds -Wno-virtual-dtor to make flags.
This warning is actually not enabled by -Wall (-Wall just enables Larry
Wall's favorite warnings). This has a kind of annoying gotcha built in
to it: protected by non virtual destructors are accepted by the
compiler. However if class that derives from such a class and does not
explicitly specify a protected and/or virtual destructor will trip the
error, for this reason it's best to always make destructors of classes
with virtual methods virtual.