fixed a bug in the memcached protocol implementation.
Removed a space from the beginning of a line in debug.hpp.
Fixed an error in the memcached protocol implementation regarding the
vesion command. We inexplicably required that the version command supply
an extra argument even though the protocol requires that the version
command take no arguments. Though this extra argument was ignored, not
supplying it was treated as an error. This was changed so that the
argument number check for the version command now expects only one
argument (the version command itself) rather than 2.

Conflicts:

	src/memcached/memcached.cc