remove load_buf_fsm_t
The only reason we kept load_buf_fsm_t around was to avoid a perf regression due
to spawning another coroutine, and with the change to the new
index_read()/block_read() interface, do_read() itself (which load_buf_fsm_t
used) spawns a coroutine, so it is junk. If spawning a coro for each load
introduces a perf problem, we're going to have to find another way to fix it.