Removed the thread parameter from the coro_t constructor.
It was not being used by anybody, and the reason it was not being used
is that it was wrong or unsafe to use, because it means we grab a
coro context from the thread we are on and move it to the thread
we are starting our coroutine on.  spawn_on_thread no longer constructed
a coro_t directly, which is why nobody was using the thread parameter
anymore.