Added metadata to store_view_t.
Because it needs to be possible to read and write metadata as part of the same transaction as a data-read, data-write or backfill operation, store_view_t now operates in terms of transactions. I have updated dummy_protocol_t to use the new transactions, but left memcached_protocol_t and redis_protocol_t as stubs. They'll have to be fixed at some point.