Split immediate-consistency code into two submodules.
Master has now been split into a master and a broadcaster; the broadcaster is responsible for creating a branch. The broadcaster, listener, replier, backfilling, and the "branch" concept are all in the `immediate_consistency/branch` subdirectory.

The concept of a master and the code for finding a master and communicating with it is in the `immediate_consistency/query/` subdirectory. This code is in a very broken state.