Add a command-line argument to prevent affinity setting
When running multiple rethinkdb servers on the same machine, they set
affinity for their threads starting with the 0'th core, which prevents
them from utilizing the unused cores with higher ordinal numbers.

The implemented solution is to not set the affinity for any of the
threads when '--no-set-affinity' command-line argument is given to the
server. This is not the best solution, but it works for the needs of
integration testing.

Also, this commit fixes a potential problem with parameter/member
variable aliasing in the linux_thread_pool_t constructor and a related
problem in the assert.