Split connectivity_t from connectivity_cluster_t
We will soon have two notions of connectivity: connectivity-layer connectivity and directory-layer connectivity. When we have made contact with a peer but not yet received their directory, we have connectivity-layer connectivity but not directory-layer connectivity. Now we have an abstract class, connectivity_t, that both of these connectivity concepts will use.