Fix compilation problems with clang.
`btree_key_buffer_t` type size was not fixed, so the compiler had to do
some ugly pointer arithmetic when accessing the fields of a class which
go after a field of `btree_key_buffer_t` type (see
`slice_keys_iterator_t` for an example). That's definitely not what we
want. Also this is a GNU extension, which is the reason why clang caught
it (lucky us!).