Add proxy nodes; change semilattice metadata to accommodate this. #583
Proxies do not appear in the machine map in the semilattice metadata and do not have reactors or
reactor drivers. They are essentially stateless and store no data on disk (apart from generating a
log file). They generate a unique machine id at startup, but this only appears in their directory
entry and is essentially unused.

A machine's port_offset is no longer part of the semilattice metadata; instead, it must be given as
a command-line option (defaulting to 0), even for established databases.