Fixed region_join for hash_region_t<key_range_t> so that it does not get its array indices backwards.
Added the test HashRegionTest.SpecialNonSquareRegionJoin which does a 2x3 region join
to check for this problem.  You would think RegionJoin2D would catch it because
it is 2x3, but one of its cells is 2x wide, or maybe it was in the wrong dimension.

This fix allows regression/issue_649.py to pass.