Adds auto key generation.
Here's how it works:
If a document is inserted without a primary key that primary key will be
filled in (discussion later of what it gets filled in with). Notice this
only works on insert. If you try to upsert then you need to specify a
key that's because presumably if you're generating keys you're expecting
them to be new so upserting would be a moot point.  Since generated keys
are exclusively upsert if you manage to get a collision the write will
fail. Since it's very unexpected for a generated key to collide (it can
only happen if there's a uuid collision or a user tries to make it
happen) a collision here throws an error.

What does the generated primary key look like?
The structure of our key is as follows:
KWPC:7f7483d4-231b-faec-091b-b0938bcacd75:4:15
1    2                                    3 4

1: A randomly generated 4 character string we don't rely on this for
uniqueness it's just there to make these a bit nicer to shard

2: The uuid of this parser each entity that can cause insertions should generate a new one of these.

3: The thread this insert was run on.

4: The number of inserts that have been run from this thread. (Incremented below.)

When pks are generated they are returned as a new field in the json
response "generated_keys". This allows a user to easily find the row
which they just inserted.

Fixes #977.