Fix a bug with thinking conflicted reads were actually errors.
The problem was that when we did a read which conflicted with a write we
would grab the data from the write rather than actually going to disk.
This is fine but we'd then checking incorrectly whether the write had
succeeded by using the same mechanism as we use for writes that do go to
disk. This involved crashing based on the value in a garbage value.

Fixes 1016.