Adds need parameters to the suggester.
The parameters are "prioritize distribution" which tells us whether to
distribute roles even if it will cause loss of service and usage which
is shared amongst suggestions and keeps track of how much we're using
each machine.
Currently prioritize distribution does nothing time to make it actually
work.

This pertains to #1232.