Removed flush_waiting_threshold configuration parameter.
It basically makes no sense.  If somebody wants an operation to wait for it to be written to disk,
they do not want to wait around 5 seconds until we feel like flushing it.  It is an artificial
parameter that affects some nonsensical benchmark.  It does not have any practical application to
our present write_durability_t scheme.