Removed .shard of the write in listener_t<>::perform_enqueued_write.
Either the change is correct, or the code was already broken.
Assertions in both the memcached and rdb_protocol write_t::shard
methods would make the code crash if the shard call could ever receive
a subset of the write's region.  The .shard call in listener.cc was
added as part of a larger change to a bunch of stuff that got things
working, not for specific discernable reason.  And it would be
obviously broken if it ever actually _did_ affect the value of the
write being passed on to the store.