Get rid of get_drain_signal method on auto_drainer_t.
It actually made no sense since the only way the signal you got back
could get pulsed was if the auto_drainer_t was destroyed but that means
you're accessing an object that could be destroyed over the course of
the function.

Instead I threaded the interruptor through to the place where it was
needed. It actually only required threading it through a single layer.