make python driver install backup scripts
rather than cycle through multiple potential python binaries with which
to run a script hoping to find the appropriate package, it makes much
more sense to install the scripts with their dependency into the
location of the python binary with which the package was installed.

this pegs the #! to the version of python used to install (whether
system version or venv) and allows the script to run in the appropriate
environment.