Fix RDBBtree.SindexEraseRange unit test
We changed how we represent an empty ungrouped stream. It used to be that an
empty ungrouped stream was represented by a map from the NULL group to an empty
`stream_t`, now it's represented by an empty map. (This is more consistent with
how we represent empty grouped streams.)

Review 1280 by @mlucy
Closes #2042