rql_test: change the sleep gap back to 0.2 after launching a rethinkdb server
The sleep gap after launching a rethinkdb server was 0.2 before, and it
was changed to 2 by commit e9e568354("Extra tweaks to the driver tests
to run on travis").

However, I just learned from issue #736 as noted by atnnn, we never use
travis anymore, and I double confirmed that while accessing
https://travis-ci.org/rethinkdb/rethinkdb; the last task was finished 7
months ago.

So, it's Okay to change the sleep gap back to 0.2 to reduce about 3
minutes testing time.

Signed-off-by: Liu Aleaxander <Aleaxander@gmail.com>