Make table interface abstract
Introduced a new abstract interface for the query language code to send queries to tables. This interface is approximately equivalent to the old table_t, except that it only represents the table itself, without any of the other data that table_t carried such as the use_outdated flag, the secondary index name, etc.

The main implementation (and currently the only implementation) is real_table_t, which translates its queries into read_t and write_t objects and passes them on to an underlying namespace_interface_t. In the future there will be another implementation for artificial tables.

I also separated out the parts of the rdb_protocol/ directory that are responsible for communicating with the B-tree and clustering code into a rdb_protocol/real_table/ sub-directory.