Make driver.py detect resunder in a more portable way
On my system, "ps -A" only prints the first word of the command line, so the word "resunder" was being cut off and driver.py believed that resunder was not running. This commit changes the arguments passed to "ps" so that it will always print the entire command line.

These changes have been reviewed by Karl Kuehn.