Get multi-value replace working.
This is not totally spec compliant yet (some renaming has to happen) and the JS
driver in atomic_get_set is experiencing some wonky nondeterminism that results
in occasional test failures.  It is not suitable for release until both these
are rectified and the code uses return_changes instead of return_vals
consistently.