Remove vector clocks
Vector clocks are now replaced with a (value, timestamp, uuid)
triple. When two triples are joined, the one with the higher
timestamp will be chosen. The UUID is randomly generated for each
triple; its only purpose is to break ties.

This commit also removes a bunch of logic for supporting /ajax,
because it's being replaced by the ReQL admin API anyway, and
removing it was less effort than updating it.