Rename `to_json` to `to_json_string`
The method named `to_json` is pseudo-magic in Ruby; the json gem
defines it on every object, and relies on a zero argument `to_json`
method existing on everything nontrivial.  This is too much of a
headache to work around in the Ruby driver alone, so we are forced to
spread the misery and rename the method.

Which is tragic, because it breaks consistency, but that’s life.