Use OpenSSL default certs automatically
This way you only have to do:

conn = r.connect(:host =>'rethink.domain.com',
                 :port => 11111,
                 :auth_key => '<auth_key>',
                 :ssl => { :ssl => true }
                 )

This looks a bit silly (and the second :ssl can actually be anything),
but :ssl is used for ssl_opts. Could change it so :ssl is just a boolean
and add a separate :ssl_opts, but this would break compatibility so
seems a bad idea. There is no other way to detect a ssl connection so
something has to be set somewhere to flag it's a ssl connection. Can't
use an empty :ssl as the code assumes that is no ssl (although suppose
could change that).

I'm also not sure if OpenSSL::X509::DEFAULT_CERT_FILE is the correct
way. Almost seems as if we should be using a default context instead of
creating a new one, although don't know if that is possible.

Note: this is just thinking out load at the moment.