Enable Server Name Indication Support in Java driver
Enabling SNI itself is easy, the tricky bit for me was modifying a test
app to work with it. For that I [applied these changes][1] on a
different branch based off the 2.3.3 driver since I had connection code
that would work with that - the recent [Jackson changes][2] here broke
my connection example.

You have to use the SSLContext option similar to [here][3]:

	private static final String DEFAULT_SSL_PROTOCOL = "TLSv1.2";

	...

	Connection conn = null;
	Connection.Builder builder = r.connection().
	    hostname(host).
	    port(port).
	    authKey(authKey);

	final SSLContext sslContext = SSLContext.getInstance(DEFAULT_SSL_PROTOCOL);
	sslContext.init(null, null, null);
	conn = builder.sslContext(sslContext).connect();

Init with all null options here is valid. It will use the default/system
trust manager for verification.

[1]: https://github.com/atomicules/rethinkdb/commit/771f34db595cd9a3131bbfe40b226fca15c49345
[2]: https://github.com/rethinkdb/rethinkdb/pull/6157
[3]: https://github.com/pires/rethinkdb-ssl-test/blob/master/src/main/java/com/github/pires/App.java#L70