edit RethinkDB license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the COPYRIGHT file so that it now contains only the
full text of the Apache license and references The Linux Foundation as
the copyright holder. The info that was formerly stored at the top of
COPYRIGHT has now been moved to the README under a new "Licensing"
section.

Collectively, these changes allow Licensee to successfully identify the
license type of RethinkDB as Apache.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>